+++ CHANGELOG.md
@@ -5,6 +5,7 @@ For more details, refer to the [documentation](https://surfquantum.github.io/yot
 ## [1.0.1] - 2025-01-16
 * added `dependabot` to update `github-actions`
 * added [zenodo](https://zenodo.org/) DOI
+* fixed for `bayesian-optimization >= 2.0.0`
 * updated dependencies
 
 ## [1.0.0] - 2024-01-31